third_party/blink/web_tests/platform/mac/external/wpt/css/css-transitions/properties-value-auto-001-expected.txt

This is a testharness.js-based test.
Found 52 tests; 0 PASS, 52 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL margin-top auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-top auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-top:2s" but got ""
FAIL margin-top auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL margin-top auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-top:2s" but got ""
FAIL margin-right auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-right auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-right:2s" but got ""
FAIL margin-right auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL margin-right auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-right:2s" but got ""
FAIL margin-bottom auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-bottom auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-bottom:2s" but got ""
FAIL margin-bottom auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL margin-bottom auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-bottom:2s" but got ""
FAIL margin-left auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-left auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-left:2s" but got ""
FAIL margin-left auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL margin-left auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-left:2s" but got ""
FAIL height auto(to) / values assert_not_equals: must not be target value after start got disallowed value "18px"
FAIL height auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "height:2s" but got ""
FAIL height auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3281px"
FAIL height auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "height:2s" but got ""
FAIL width auto(to) / values assert_not_equals: must not be target value after start got disallowed value "100px"
FAIL width auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "width:2s" but got ""
FAIL width auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3281px"
FAIL width auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "width:2s" but got ""
FAIL clip auto(to) / values assert_not_equals: must not be target value after start got disallowed value "auto"
FAIL clip auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "clip:2s" but got ""
FAIL clip auto(from) / values assert_not_equals: must not be target value after start got disallowed value "rect(15px, 15px, 5px, 5px)"
FAIL clip auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "clip:2s" but got ""
FAIL marker-offset auto(to) / values assert_not_equals: initial and target values may not match got disallowed value ""
FAIL marker-offset auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "marker-offset:2s" but got ""
FAIL marker-offset auto(from) / values assert_not_equals: initial and target values may not match got disallowed value ""
FAIL marker-offset auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "marker-offset:2s" but got ""
FAIL top auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL top auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "top:2s" but got ""
FAIL top auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL top auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "top:2s" but got ""
FAIL right auto(to) / values assert_not_equals: must not be target value after start got disallowed value "22.4688px"
FAIL right auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "right:2s" but got ""
FAIL right auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL right auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "right:2s" but got ""
FAIL left auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL left auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "left:2s" but got ""
FAIL left auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL left auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "left:2s" but got ""
FAIL bottom auto(to) / values assert_not_equals: must not be target value after start got disallowed value "82px"
FAIL bottom auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "bottom:2s" but got ""
FAIL bottom auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.3333px"
FAIL bottom auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "bottom:2s" but got ""
FAIL z-index auto(to) / values assert_not_equals: must not be target value after start got disallowed value "auto"
FAIL z-index auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "z-index:2s" but got ""
FAIL z-index auto(from) / values assert_not_equals: must not be target value after start got disallowed value "10"
FAIL z-index auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "z-index:2s" but got ""
Harness: the test ran to completion.
